Move RowIterator to execution package
